Struct eagre_ecs::system::System

pub struct System {
    // some fields omitted
}

An Entity-Component System

Example

// Those two traits are required for components
#[derive(Debug, Clone)]
struct MyComponent(i32);

let mut system = System::new();
let entity = system.new_entity();
system.add(entity, MyComponent(16)).unwrap();
system.run::<MyComponent, _>(|sys: &System, ent: Entity| {
    println!("Entity {} has {:?}", ent, sys.borrow::<MyComponent>(ent).unwrap());
}).unwrap();

Methods

impl System

fn new() -> System

Create a new Entity-Component System

fn new_entity(&mut self) -> Entity

Create a new entity

fn remove_entity(&mut self, ent: Entity) -> EcsResult<()>

Destroy an entity and all it's components

Returns an Error if the Entity does not exist

fn num_entities(&self) -> usize

Return the number of entities, this system is currently containing

fn add<T: Component>(&mut self, ent: Entity, comp: T) -> EcsResult<()>

Add a component to an entity

fn set<T: Component>(&mut self, ent: Entity, comp: T) -> EcsResult<Option<T>>

Set a component's value for an entity

Returns a previous value, if one exists

fn remove<T: Component>(&mut self, ent: Entity) -> EcsResult<()>

Remove a component from an entity

fn borrow<T: Component>(&self, ent: Entity) -> EcsResult<&T>

Borrow a component of an entity

fn borrow_mut<T: Component>(&mut self, ent: Entity) -> EcsResult<&mut T>

Borrow a component of an entity mutably

fn get<T: Component>(&self, ent: Entity) -> EcsResult<T>

Get a copy of a component

fn has<T: Component>(&self, ent: Entity) -> bool

Test, if an entity has a component

fn run<T: Component, F: FnMut(&System, Entity)>(&self, f: F) -> EcsResult<()>

Run a procedure for every entity with a component

fn run_mut<T: Component, F: FnMut(&mut System, Entity)>(&mut self, f: F) -> EcsResult<()>

Run a procedure for every entity with a component, with changes to the system being allowed